stack new kmonad  --bare new-template
This is the straight output of `stack new kmonad  --bare new-template`
with no parameters. We chose this particular template[1] because it
gives us the framework for a library, unit tests and a main program;
we probably don't need the last one but we'll remove it when we
confirm that.

This also includes licence and other files that we will clean up
in the next commit.

[1]: https://github.com/commercialhaskell/stack-templates/wiki