src/LHTutorial2: Build code from §2.4 of the Liquid Haskell tutorial
XXX Liquid Haskell generates a bunch of files under .liquid/ in the
source directory; obviously this belongs under .build/ in the project
directory instead. But we don't know how to fix that right now, so in
the interests of moving forward, we ignore that and put a pin in this.

Liquid Haskell requires an SMT solver, so Test now checks to see if
`z3` (the Z3 theorem prover[1]) is installed and attempts to install
it (Debian-style) if not. It will still fail on non-Debian-based
systems but we print a hint about what we're trying to do that any
Linux user should understand. We considered just asking the user to
install it herself, but since the Haskell Stack installer is already
doing `sudo apt install` stuff anyway, we might as well just carry on
with that.

stack.yaml has been completely replaced, removing a lot of comment
verbiage and setting things up properly for Liquid Haskell, which
wants a specific point version of GHC (9.8.2 because the 9.8.4 default
for Stackage LTS 23.24 fails builds). There are LH builds for GHC 9.10
and 9.12, but we probably just want to stick with the minor release of
the latest Stackage LTS.

[1]: https://github.com/Z3Prover/z3