lhver/: Configs for LiquidHaskell 0.9.6.3, 0.9.8.1, 0.9.8.2
0.9.8.2 is what used to be in the top-level stack.yaml, and we use
that via a symlink. 0.9.6.3 and 0.9.8.1 are copied from the sample
files in the `lh-plugin-demo` repo (see lhver/README for more).

This sets us up to try different versions of Liquid Haskell (which we want
in the long run anyway) and, right now, lets us work from 0.9.8.1, for
which we have an example, and then see if we can cons up a 0.9.8.2 that
will match it.

All of these work with the current (extremely simple) code, but none of
them yet work with the symbolic execution example that requires
Language.Haskell.Liquid.ProofCombinators.