WIP src/Fib: Make LH output verbose
Adding an OPTIONS_GHC pragma:

    {-# OPTIONS_GHC -fplugin-opt=LiquidHaskell:--verbose #-}

is the wrong way to do this. we want to be able to specify a command line
option to `Test` to dump this to a file in the build directory, so we don't
need to change code to turn it on and off. (Or possibly just always dumping
it would be fine, if it doesn't slow things down.)