WIP src/Fib: Break proof so we can debug the output
By changing the '9' to an '8' we make the proof invalid; the next step
is to look at the verbose output and figure out more about what it's
passing to the SMT solver. (Search for 'SMT' in the verbose output.)
This is going to take some work for us to parse and figure out.