WIP Imp: Remainder of syntax, sample program, improved docs
XXX As develop this, we want an easy way to `show sum_imp` (or other
things) after a build. (We don't really need to bother with a unit test for
what show produces because if the `show` works, the type system has taken
care of almost all the rest.) Can we do this by having Test feed an
expression into ghci? The more traditional way is to run ghci and manually
:reload and type in your expression again every time, which is tedious.

TODO in program:
- Confirm the comment for mkStmt is correct.
- Clean up sum_imp, especially to remove parens and make use of precedence.

TODO document here:
- Head comment, refrences
- StandaloneDeriving
- 'KImp'
- Remainder of abstract syntax.
- Sample program from tutorial
- mkStmt