**/*.yaml: Use consistent two-space indentation, for lists especially
YAML allows one not to indent or to indent a list giving items for an
option, e.g.:

    shorter:            shorter:
    - one                 - one
    longer:             longer:
    - two                 - two
    - three               - three

The latter is slightly easier to read, so we go with that. (And
it also helps keep things from getting even worse in certain other
situations, such as some examples, given at
<https://stackoverflow.com/a/71968274/107294>).

We go with two column indents rather than the usual four because these
files tend to have longish lines when comments are added, and due to
the syntax and contents it can be quickly read without confusion even
if you don't grasp the exact indent level.