CImp: Extract Store; State now includes Prog
The particular way we handle execution state such as our location in
the program is dependent on the semantics; this can be done in vastly
different ways. (In this case, we've chosen to remove each line from
the head of the Prog as it's executed, which is fine since we have no
looping constructs.)

But in all cases it makes sense to have however you do it be embedded
in the State, which then becomes dependent on the particular
semantics, just as the store (if you have one, or more) is.

So we extract the store (often called σ) from the State as its own
thing and the State is repurposed to hold *all* execution information,
including program "location" etc.