app/Benchmark: Run KImp.sum_imp with 10e6 iterations
Note that we have to be careful not to let Haskell get too lazy and
just skip most of the actual evaluation. (Though this does produce
excellent results, optimising the algorithm to O(1).)

At ten million iterations (same as used in bench/Bench for the K
version), the Haskell version is about 50% slower.