package,Bench: Enable +RTS for all compiled programs; show stats in benchmark run
We probably want later to add `-rtsopts` only to the benchmark
programs we compile, if we discover that there's any performance hit
from this, but it's not clear that it is in which case having it on
everything we build is really nice.

We add `+RTS -sstdout` to the run in bench/Bench so that we see the
full runtime stats: mainly GC information, though it also handily
gives us timings as well. (We still keep the /bin/time output, though
so that we can be sure that we're comparing the exact same thing
with the K benchmark run.