KImp: strictify AExp = Int !Integer
This makes the GC time go away, but we obviously don't want everybody
to be having to strictify stuff internally in their Semantics, and that's
not even general anyway, depending on what the program does.

This just shows that doing this for `sum_imp` happens to work, but we
are going to work on instead having `eval` do a `deepseq` on the
state with every rule application, so that it doesn't really matter
how lazy the Semantics are except in pathological situations.