KImp: Unroll eval.applyFirstMatchingRule
Recursive calls to `eval` unrolled by an `orElse` binary function.
See bench/Notes.md §3 for details.

`orElse` is a monadic-type operator that encodes a "first match"
strategy for the composed rewrite rules.

This seems to want to turn into a proper monad, or rather set of
monads each of which can implement a different evaluation strategy
(first-match, branch-all-in-tree/multiverse, etc.)

The output of ghc-core on this commit of src/KImp.hs needs to be examined
and compared with the ghc-core output for the previous commit so we can
better understand what optimisations GHC is applying here.