KImp: Manually sqaush all rewrites in to a single function
This allows GHC to do its super-duper optimisation on dealing with a
large set of pattern matches. (It will do this for all patterns of one
function, but can't do it for a bunch of separately applied functions
with similar patterns.)

This brings our runtime down massively, from over 11 seconds in the
previous commit to about 4.2-4.5 seconds in this commit. (This was
checked with a couple different orderings of the patterns and that
seemed to make no difference, as expected.) This is now about twice
as fast as the K -O3 version, which runs in 9s or so.