package,app/OptTest: All app/* modules become 'module Main'
The `other-modules: []` line is the key to ensuring that these don't
produce mysterious build errors: the problem appears to be that everything
in `app/*.hs` was being built together for every exectuable in `app/`,
which of course caused collisions if you had more than one module named
`Main`.

Using `module Main` for all of these allows us to get rid of the
`ghc-options: -main-is OptTest.main` lines. What the module actually
is is still clear enough from the filename.