app/Hello: Print hello and arguments
This serves as a quick example of how to read command line arguments.
Additionally, being our second executable, it demonstrates:

To handle having multiple modules in a directory with the same name
(`module Main` in this case), we must explicitly specify an `other-modules:
[]` for the build of that object because the default is all the modules in
`source-dirs:`, which would cause it to collected with the other `module
Main` modules when linked. See:
• https://github.com/sol/hpack?tab=readme-ov-file#executable-fields