doc/*SO-post → doc/discussion/*SO-redundant-*: Separate dir for Q&A posts
doc/discussion/ is copies of posts to Q&A sites (e.g., Stack Overflow,
maybe Reddit). We use this just for larger and more detailed posts that
require significant editing and tuning before posting, which we do
here and then paste a copy.

The separate subdir not only makes for less clutter given that we're adding
additional posts, but also lets us use longer and more descriptive names
for the post files.