Add unit tests for KTutImp and CleanImp
These are placed in separate modules, KTutImpTest and CleanImpTest,
so that dependencies of the main semantics modules do not depend on the
testing framework. Unfortunately, any dependencies on the kmonad library will pull in
tasty-hunit and tasty as dependencies.

For now, however, this lets us get moving with regular tests.