Tarnyq: Extract common code from src/*Imp.hs'; foldr for eval
The Tarnyq module now contains the core of the semantics/eval
stuff. KTutImp and CleanImp now both use this. It took some
tricks with INLINE to keep the speed the same.

The change to eval to use foldr lets us go back to representing
the semantics as a list of rewrites. Surprisingly, we do not see
the performance drop Nishant expected. Possibly previous
experiments were using a different version of fold, even a
hand-written one instead of a library one.

Note that we do see a performance degradation if we use the
version of `eval` that used to be CleanImp, which walks through
the list itself. (This is what was used in the §2 benchmark
timings.)