NonDet: Toy language for testing/playing with non-determinism
We can see that it is non-deterministic by reversing the order of the
rules in the semantics.

The evaluation algorithm does a depth-first evaulation, though in
a somewhat efficient way. (In particular, it's not tail recursive and
it uses `++`/list concatenation.)

This also changes 'Semantics a' to '[Rewrite a]' because the
abstraction to 'Semantics' obscures things rather than offering extra
clarity.