stack.yaml,lhver/0.9.10.1: Switch to GHC 9.10.1 from Stack lts-24.7
Stackage resolver lts-24.0, supporting GHC 9.10, was released on
2025-07-13, and Liquid Haskell now supports 9.10.1, so update to this
more recent version. (Note that we have to back down from the default
9.10.2 because LH does not support that version.)

GHC 9.10 also happens to give us about a 10% speed increase if we
use `-fllvm`.

LH also supports GHC 9.12.2 [1], but Stack does not support that in a
released version yet, just in the nightly.

[1]: https://ucsd-progsys.github.io/liquidhaskell/install/#installing-liquidhaskell