lhver/0.9.10.1: Note that LH seems to work fine with 9.10.2, as well
This was tested by simply changing the compiler for 9.10.1 to 9.10.2;
it built and all tests passed.

That said, we're not changing this now since when we switch away from
Stack we won't be using this file; the build will have a different,
non-`stack.yaml` way of determining which GHC version to use.