Test: Add $TOOLBIN, install_hpack(), and run hpack
This now generates Tarnyq.cabal, meaning that we can now remove Stack
and do our build with just `cabal` (modulo configuring it correctly,
particularly for the compiler version).

This works only on Linux x86_64 at the moment because of the way we
bootstrap `hpack`. There are various options for that. Ideally we'll
do a proper bootstrap from Hackage to build a local binary, but there's
also the option of just using an OS package (Debian 12 has 0.34 with
basically no deps, and it's in Brew as well). We'll figure that out
once we get our GHC version selection working.