Test find_cabal(): Find a `cabal` executable
This uses what the search path gives us, if anything, and otherwise
finds one installed by `ghcup`. Possibly further searching could be
done, but there are no obvious other places to look. (If there are
Stack-installed tools, these don't include `cabal` because Stack has
that built-in to itself.)

We expect that users concerned about path polution will not have
`cabal` in their path; we should provide a way to easily give these
users an alias or shell function to run the `cabal` we found.

Cabal is not yet used; we still build with Stack.