Test: Add -G option to set specific GHC and find_ghc() to just find one
This ensures that a `cabal.project.local` file exists and has a
`with-compiler` option in it. If the user's not explicitly asked to
use a specific compiler, find_ghc() will search for a ghcup-installed
one from a list of preferred versions (failing if none of the
preferred versions are available).

The -G option takes a path to a GHC executable and forces the use of
exactly that version. (No warnings are given for versions that may not
work.)

This still does not yet actually use cabal and that GHC to build.