Test: build/run `ghc-core`, which makes the `-g` option work again
The ghc-core run is moved ahead of running ghci, since the former is
part of the build.

`ghc-core` appears to try to run just `ghc` (or something—exactly what
has not been confirmed), rather than the executable given in the
'with-compiler:' field of cabal.project.local. So we just manually set
the --with-compiler option to the contents of that field.