Test: Suggested minimum cabal version is 3.12, and (bad) explanation
It was 3.11, which is basically the same thing, but we were going
to bring it down to 3.10 after testing. But we see now that the
"Nix-style standbox" stuff came out in 3.12, and rebuilding with
3.10 actually rebuilds everything, so the object directories are
somewhere different with that version. (I think that 3.12 brought
in global versioned object dirs, or something like that.)

We really should explain this better, but it's good enough for
the moment, because it's just a warning anyway.