Test: Re-order dev tool installs/builds more sensibly
We always start first with cabal and then ghc, since that gives us the
tools we'll need to boostrap a compile of HPack when we do that. (We
need to set the GHC version "manually" in cabal.project.local because
we may not have a *.cabal file(s) that's necessary to do a `cabal
configure`. It shouldn't need this, but it does.)

Z3 moves to the end simply because it's not really related to the
earlier stuff; it's used during compile but only for Liquid Haskell.