Test with_compiler(): Print only last with-compiler: line
We may have multiple `with-compiler: …` lines in the file because of
the slightly hacky (but easy) way we set it, or just because someone
happened to manually edit it and leave it with multiple with_compiler
lines.

Print only the last one, which is the one that cabal will use.