Test: Consolidate sysem package installs and add ones that were missing
The install_syspkgs() function now deals with checking for all system
packages that we require, and attempting an install of any that are
missing. (Comments in the list of packages in this function indicate
for what they're needed.)

This adds a number of packages that developers usually already have
installed, but were exposed by running the build in a fresh Debian 12
Docker container.