Test: Check length of array ghci_exprs directly, instead of expanding
We were using the ${ghci_exprs[@]} expansion/substitution of values in
the array and checking to see if the result was zero length, which is
fine in Bash 5.2, but with `set -u` generates an error 'unbound
variable' in Bash 3.2. Arguably the latter is really the correct
behaviour (because if you're expanding to give the value of index 0,
index 1, etc., and index 0 is unset, the whole thing should be
considered unset).

Anyway we are now directly checking the length rather than expanding
and checking the length of the expansion, which is more clear anyway.