Test,cabal.project: ncabal(); Enable use-pkg-config for pcre-light
The new ncabal() function in Test runs the `cabal` selected for this
project, but does so in a current working directory that has no files
in it (in particular, no `cabal.project*` files) and with the project
dir set to that directory so that cabal will not search upwards for
`cabal.project*` files.

This fixes the problem where if any `cabal.project*` files are found,
cabal often must have one or more `*.cabal` files available as well or
it will exit with an error, even when doing something that doesn't
read any `*.cabal` files. To deal with the lack of a toolchain setting
that we'd normally get from `cabal.project.local`, we instead provide
the one from our "real" cabal.project.local at the command line.

Further details are in the comments.