XXX cabal.project: Enable use-pkg-config for pcre-light
XXX need to explain all this, especially new ncabal() etc.

See comments for details.

XXX OBSOLETE This breaks our clean build because now that we have a
`cabal.project` file, Cabal attempts to use it when we run `cabal update`
and it fails because we've not yet generated the Tarniq.cabal file, but
`packages: .` requires us to have at least one .cabal file in our project
dir.