Re-implement rewriting using a monadic interface...
TODO Document summarized
TODO Maybe this commit message could be turned into documentation

The names of the modules, types, and classes used here, are modelled after
the [MTL]'s State monad.

We extract an abstract interface, MonadRewrite, that provides an the
tools necessary for rewriting---getting the LHS, matching on it, and
constructing the RHS. Language semantics are defined using this
interface---each rule has a type signature
`MonadRewrite m State => m ()`, where `m` is a parameter that may be
instantiated by any monad implementing the MonadRewrite interface, and
the `State` parameter is the languages State type. We no longer
lift{AExp,BExp,K,Store}, etc, because helpers such as `getK` and `putK`
get us most of the convenience they offered.

`get{AExp,BExp,Stmts}` and `put{AExp,BExp,Stmts}` were not implemented
because the `put` variant would be partial. I didn't like the assymetry
and they do not give us much over `getK`.

We also define a new abstract interface ProgramIO that allows printing
to the console.

We provide three instances of MonadRewrite:

1.  `Basic.Rewrite` allows rewriting only over simple State ADTs. It
    offers evaluation for both the one-path, and all-path variants.

    These are currently defined as two different functions, but could
    easily be split out into two implementations by using List, instead
    of Maybe, as the inner monad. That is, use:

        `{ getFun :: s -> [] (a, s) }`

2.  `IO.RewritePure` allows rewriting with *mocked* IO, stored by
    wrapping `Basic.Rewrite` in a pair. To use this IO interface, a
    language semantics must also use the `ProgramIO` constraint over the
    moand:

        (ProgramIO r, MonadRewrite r State) => [r ()]

3.  `IO.RewriteIO` allows rewriting with *actual* IO, by using
    `IO (Maybe _)` as the inner monad. Again, the `ProgramIO` constraint
    must be used.