bench/Bench: Explicitly use $TOOLBIN/cabal instead of relying on $PATH
We set $PATH to have $TOOLBIN at the front as a backstop, since
otherwise we might silently run different versions of cabal, ghc-core,
etc. that happen to be in $PATH, but it's best to explicitly show that
we're using the $TOOLBIN version anyway to make it clear to the reader
that it's not just any old version of these programs.