src/ → semantics/: (Step 1) Move all semantics and programs
This is the first step in splitting out our semantics and programs
from the core Tarnyq libary. The core library and various supporting
tools remain under src/; the semantics and programs are moved to the
semantics/ subdir; and we will further separate the programs from the
semantics to move them to the program/ dir.