exe/Benchmark → HandRun: Rename program
This executable is no longer used just for running the benchmark, but
also for running arbitrary hand-coded functions. (We pattern match
args to determine which one to run.) We rename it to more clearly
state what this does, and also add a note that this will be replaced
at some point by an executable  that will take an arbitrary evaluator,
semantics and program and run it, so that we don't have to continue
hand-coding the calls that do this.