binech/Bench: Supply imp/sum/count args to hand-run
This directly supplies to `hand-run` the parameters to match the
function that calls 'eval_imp' with the parametrized sum_imp program
with the correct count, rather than relying on the default
no-parameters call to `hand-run` to do this.

This not only moves the parameters closer to the benchmark code, so
it's more clear what it's doing, but allows us to remove the default
no-parameter function call in `hand-run` that duplicated what we can
do with the `imp sum N` args.