program/KTutImp_Program: Extract sample/test 'imp' programs from KTutImp
The long-term intention here is that you can create any
`program/S_N.hs` file containing programs in semantics _S_ where
_N_ is an arbitrary name for the group of programs in that file.
Each program should be a top-level definition that's exported
from the module.

These should eventually be auto-discovered in a way that will let you,
from the command line, specify a semantics, an evaluator, and a
program name (and possibly a group name, if needed to distinguish
programs with the same name from different groups) and run that
program with that evaluator.