exe/ShowExports: `show-exports FILE.hs` lists explicit exports
This gets us a little way towards being able to create a "generic
evaluator" command-line program that can run arbitrary code without us
having to hand-code the imports and names into the command-line
program. Given a source file, it prints the module name and the
explicit exports. (If no export list is given, it does not, as Haskell
does, assume that all top-level definitions are exported, though we
could do further parsing to implement this.)

This uses the haskell-src-exts package, which is a fairly lightweight
parser that doesn't handle preprocessors such as Template Haskell. for
that, we'd probably need to move to parsing using ghc-lib-parser or
similar. But since we also need to parse only the head information of the
module (in fact, just the module name and export list, not even the
imports), we might be able to get around the need for something more
complex by using the haskell-src-exts's lazy parser and just letting it
stop at anything it can't handle after the module name and export list.