Rewrite.IO: Cleanup interface for IO by using Integers rather than Strings
The semantics for reading Integers is better defined than for reading Strings,
since it is debatable whether we want to read a line, or a word or as much
is available in the stream. We will probably want to split this into
two interfaces:

1.  An interface for reading arbitrary types that implement Read,
2.  A POSIX abstraction, for more serious languages.

This also moves to the cleaner readMaybe interface, rather than crashing
when parsing as Ints doesn't work.